Refactor: CardListPresProps 타입 분리 및 타입 파일 관리 구조 개선
