+++ views/card-list/types.ts
@@ -15,5 +15,8 @@ export type ViewType = 'card' | 'list';
 export type CardListPresProps = {
   viewType: ViewType;
   setViewType: (viewType: ViewType) => void;
+  sort: 'latest' | 'popular';
+  setSort: (sort: 'latest' | 'popular') => void;
   items: CardData[];
+  sortOptions: { label: string; value: string }[];
 };
